demoapp/tests/requirements/test_requirements.txt
pytz
nh3
django-ckeditor
psycopg2-binary
pytest
pytest-django
celery
lxml
pytest-django
pytest-mock
pytest-cov
redis
requests